test: 컨트롤러 테스트 문서화
